requirements.txt
pillow
exiftool
